Fixed the integration + tests.
The integration is now extended so that results from both plugins can be
matched. This is used to give GRMT results the same title + authors as
the corresponding GR results, because otherwise they may not be merged.

The tests now mock the browser, so they no longer depend on the
goodreads server.