Attempt to avoid deadlocks
There is now a timeout on how long we will wait for data from the
Goodreads plugin. If it takes longer than this timeout, either due to
slowness in the network or due to a bug in our code, we will stop
waiting and carry on with what we have.