Quickly detect if Goodreads is part of current run
If the Goodreads plugin is installed/enabled the integration is detected
as being present, regardless of whether it is actually used in the
current identify call. This causes this plugin to wait until the timeout
hits before fallbacking.

This change makes it so that the Goodreads plugin receiving a start
signal is detected, and if this doesn't occur within a second or so of
this plugin receiving a start signal it is assumed the Goodreads plugin
will not be part of this identify, and we proceed as if the integration
was disabled.